balsn-ctf-2019/images-and-words/docker/zeus/app_scaffold/requirements.txt
aiofiles==0.4.0
Click==7.0
gunicorn==19.9.0
h11==0.8.1
httptools==0.0.13
pypng==0.0.20
python-multipart==0.0.5
six==1.12.0
starlette==0.12.9
uvicorn==0.9.0
uvloop==0.13.0
websockets==8.0.2
